+++ src/worker/libraries/common/v30/bslDefender.brs
