+++ src/worker/libraries/models.csv
@@ -40,6 +40,7 @@ N1101,Roku HD-XR Classic,STB,directfb,hd,D0
 4660X,Roku Ultra (2017),STB,opengl,fhd,YJ
 4670X,Roku Ultra (2019),STB,opengl,fhd,YJ
 4800X,Roku Ultra (2020),STB,opengl,fhd,X0
+4850X,Roku Ultra (2024),STB,opengl,fhd,X0
 5000X,Roku TV (MIPS),TV,directfb,hd,2N
 6000X,4K Roku TV,TV,opengl,fhd,YW
 7000X,4K Roku TV (TCL),TV,opengl,fhd,YN
